Fix blending by moving it to show-time rather than fetch-time
Previously I implemented video blending at fetch time, but this was inaccurate (you could fetch a video right after showing one, but not actually show a video for another minute).

This also fixes `isAvailable` by checking interstitial blending rules for that session. This has the side effect of requiring /mediate to be called, which is good (previously we said an ad was available before /mediate was called, so if you called show you would have to wait for /mediate to finish).