Ads caching layer to vast video & requires cached video to allow play. Fixes bug in HZVideoAdModel also.
Also removes SKVAST reachability code and references - not necessary for us.
Bug in HZVideoAdModel: URLForVideo returned the cached file url if the file exists, but the file will always exist once the cache download has begun. Fixed it to only return the cached URL if the video has finished downloading.