Remove the AppLovin didLoad/didFail abstract listener callbacks.
These callbacks are only sent for the instances of the ads that we create; they are irrelevant/useless to a dev. Even if we load an incentivized ad, the `didLoad` callback is only applicable for *that instance* of the incentivized ad. The developer isn't able to show an incentivized ad at that point.