takes the load manager aware of segmentation.
This will allow it to continue lazy loading networks until a network has an ad that will be allowed by segmentation.
There is also a special case added to segmentation - it can tell if a segment will never allow an ad right now for a given adapter. This way, we don't lazy load all networks and fetch them, only to determine later they won't be able to show anything.