After 15 seconds, allow another ad to be shown
We've had a number of issues where a 3rd party network never sends a "dismiss" callback for one reason or another. Our mediation layer would believe an ad was still showing, so it wouldn't allow another ad to show.

To fix this, after 15 seconds we assume an ad is "stale". We'll still take callbacks from that ad network (so if an ad is like 60 seconds or something, our show/hide will still be balanced correctly), but we'll allow another ad to be shown past that point.