moves interstitial video logic out to separate class, and changes fetch logic re: interstitial video
videos should not be fetched if interstitial video is completely disabled. However, we do want to fetch a video creative when fetching HZInterstitialAd even if the current interstitial video interval will prevent us from using the creative at the moment, so that if the timeout defined by the interval expires between the fetch and a call to show we'll be able to show the fetched video. But in this scenario, we won't send fetch success for video fetches - this prevents the odd situation where fetch reports success for a video-only network because a video is ready, but a call to `isAvailable`/`show` will not see an ad ready since interstitial video is on a timeout.